requirements.txt
numpy==1.20.0
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
speedtest-cli==2.1.2
